ABSENTEEISM DURING WORK HOURS
The exercise will address Absenteeism at company during work hour.purpose of the business exercise will be to explore whether a person presenting certain characteristics is expected to be away from work at some point in time or not.We will use module pandas and give the name pd so that every time we need this module while coding we will just type pd for short instead of writing pandas more technically speaking. We want to use precisely this library because it possesses various tools for handling data in a tabular format . We will start working on the ‘Absenteeism_data.csv’ file and take it to a usable state in a machine learning algorithm.